feat: manage modals global
- move SignUp Component
- creat modalHandleStore with zustand
