+++ client/src/modals/SignUp/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUp } from "./SignUp";
