+++ client/src/pages/Home/Home.tsx
@@ -1,8 +1,7 @@
 import styled from "styled-components";
 import { useParams } from "react-router-dom";
-import { useState } from "react";
 import { Container } from "../../components";
-import { HeaderElement, MainElement, FooterElement, SignUp } from "./components";
+import { HeaderElement, MainElement, FooterElement } from "./components";
 
 const LayoutStyle = styled.div`
   display: flex;
@@ -29,24 +28,20 @@ const S = {
 
 const Home = () => {
   const { paramRoomCode } = useParams();
-  const [modal, setModal] = useState<boolean>(false);
   return (
-    <>
-      {modal === true ? <SignUp modal={modal} setModal={setModal} /> : null}
-      <Container>
-        <>
-          <S.HeaderLayout>
-            <HeaderElement />
-          </S.HeaderLayout>
-          <S.MainLayout>
-            <MainElement modal={modal} setModal={setModal} paramRoomCode={paramRoomCode} />
-          </S.MainLayout>
-          <S.FooterLayout>
-            <FooterElement />
-          </S.FooterLayout>
-        </>
-      </Container>
-    </>
+    <Container>
+      <>
+        <S.HeaderLayout>
+          <HeaderElement />
+        </S.HeaderLayout>
+        <S.MainLayout>
+          <MainElement paramRoomCode={paramRoomCode} />
+        </S.MainLayout>
+        <S.FooterLayout>
+          <FooterElement />
+        </S.FooterLayout>
+      </>
+    </Container>
   );
 };
