+++ client/src/pages/Home/components/index.ts
@@ -1,4 +1,3 @@
 export { default as HeaderElement } from "./HeaderElement";
 export { default as MainElement } from "./MainElement";
 export { default as FooterElement } from "./FooterElement";
-export { default as SignUp } from "./SignUp";
