+++ client/src/pages/Reset/index.ts
@@ -1,0 +1,1 @@
+export { default as Reset } from "./Reset";
