+++ client/src/store/index.ts
@@ -1,3 +1,4 @@
 export { default as userStore } from "./userStore";
 export { default as socketStore } from "./socketStore";
 export { default as lobbyStore } from "./lobbyStore";
+export { default as modalHandleStore } from "./modalHandleStore";
